data_server/requirements.txt
django
djangorestframework
django-environ
django-cors-headers
psycopg2-binary
firebase-admin